+++ spec/spec_helper.rb
@@ -26,6 +26,7 @@ require "rspec"
 require "rdkafka"
 require "timeout"
 require "securerandom"
+require "digest"
 
 # Load support modules
 require_relative "support/kafka_config_helpers"